+++ nab-hibernate-persistence-provider/src/main/java/ru/hh/nab/hibernate/NabHibernateCommonConfig.java
@@ -18,10 +18,10 @@ import ru.hh.nab.hibernate.events.EventListenerRegistryPropagator;
 import ru.hh.nab.hibernate.service.NabServiceContributor;
 import ru.hh.nab.hibernate.service.ServiceSupplier;
 import ru.hh.nab.hibernate.transaction.DataSourcesReadyTarget;
-import ru.hh.nab.hibernate.transaction.ExecuteOnDataSourceAspect;
 import ru.hh.nab.hibernate.transaction.ExecuteOnDataSourceBeanPostProcessor;
-import ru.hh.nab.hibernate.transaction.ExecuteOnDataSourceTransactionCallbackFactory;
 import ru.hh.nab.hibernate.transaction.ExecuteOnDataSourceTransactionCallbackFactoryImpl;
+import ru.hh.nab.jdbc.aspect.ExecuteOnDataSourceAspect;
+import ru.hh.nab.jdbc.aspect.ExecuteOnDataSourceTransactionCallbackFactory;
 import ru.hh.nab.jdbc.routing.RoutingDataSourceFactory;
 import ru.hh.nab.jdbc.transaction.DataSourceContextTransactionManager;
 import ru.hh.nab.jdbc.transaction.TransactionalScope;
