+++ nab-jdbc/pom.xml
@@ -31,6 +31,16 @@
             <artifactId>spring-jdbc</artifactId>
         </dependency>
 
+        <!-- AspectJ -->
+        <dependency>
+            <groupId>org.aspectj</groupId>
+            <artifactId>aspectjrt</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.aspectj</groupId>
+            <artifactId>aspectjweaver</artifactId>
+        </dependency>
+
         <!--Web-->
         <dependency>
             <groupId>jakarta.servlet</groupId>
